\name{armandopack-package}
\alias{armandopack-package}
\alias{armandopack}
\docType{package}
\title{
\packageTitle{armandopack}
}
\description{
\packageDescription{armandopack}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{armandopack}
\packageIndices{armandopack}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{armandopack}

Maintainer: \packageMaintainer{armandopack}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
suma.dos.numeros <- function(x,y) {
   x + y
}
suma.dos.numeros (5,6)

resta.dos.numeros <- function(x,y) {
  x - y
}
resta.dos.numeros(5,6)

multiplica.dos.numeros <- function(x,y) {
  x * y
}
multiplica.dos.numeros(5,6)
}
